+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Noteable changes of Potori will be documented here. For [Status Data](./src/data
 ## [0.7.12](https://github.com/lucka-me/potori/releases/tag/0.7.12) - 2020-TBA
 - TBA
 
+### Added
+- Double click app bar to back to top
+
 ### Changed
 - Minor UI improved